Crate bitcoin_internals

Rust Bitcoin Internal

This crate is only meant to be used internally by crates in the rust-bitcoin ecosystem.

Modules

error

Error

hex

Helpers for encoding bytes as hex strings.

macros

Various macros used by the Rust Bitcoin ecosystem.

serde

Contains extensions of serde and internal reexports.

Macros

const_assert

Asserts a boolean expression at compile time.

debug_from_display

Implements Debug by calling through to Display.

fmt_hex_exact

Format known-length array as hex.

impl_array_newtype

Implements standard array methods for a given wrapper type.

impl_parse

Implements conversions from various string types.

impl_parse_and_serde

Implements conversions from various string types as well as serde (de)serialization.

impl_std_error

Impls std::error::Error for the specified type with appropriate attributes, possibly returning source.

parse_error_type

Creates an error type intended for string parsing errors.

write_err

Formats error.