Struct bitcoin_internals::hex::buf_encoder::OutBytes

pub struct OutBytes(/* private fields */);

A byte buffer that can only be written-into.

You shouldn’t concern yourself with this, just call BufEncoder::new with your array.

This prepares the API for potential future support of [MaybeUninit<u8>]. We don’t want to use unsafe until it’s proven to be needed but if it does we have an easy, compatible upgrade option.

Warning: repr(transparent) is an internal implementation detail and must not be relied on!