Trait bitcoin_internals::hex::display::DisplayHex

pub trait DisplayHex: Copy + IsRef {
    type Display: LowerHex + UpperHex;

    // Required method
    fn as_hex(self) -> Self::Display;

    // Provided method
    fn hex_reserve_suggestion(self) -> usize { ... }
}

Extension trait for types that can be displayed as hex.

Types that have a single, obvious text representation being hex should not implement this trait and simply implement Display instead.

This trait should be generally implemented for references only. We would prefer to use GAT but that is beyond our MSRV. As a lint we require the IsRef trait which is implemented for all references.

Required Associated Types

type Display: LowerHex + UpperHex

The type providing fmt::Display implementation.

This is usually a wrapper type holding a reference to Self.

Required Methods

fn as_hex(self) -> Self::Display

Display Self as a continuous sequence of ASCII hex chars.

Provided Methods

fn hex_reserve_suggestion(self) -> usize

Hints how much bytes to reserve when creating a String.

Implementors that know the number of produced bytes upfront should override this. Defaults to 0.

Object Safety

This trait is not object safe.